requirements.txt
datasets
dill
huggingface_hub
huggingface_hub[cli]
jsonlines
openai
PyYAML
tqdm
